+++ src/database/migrations/2016_08_07_152214_add_table_cms_privileges_roles.php
@@ -13,15 +13,15 @@ class AddTableCmsPrivilegesRoles extends Migration
     public function up()
     {
         Schema::create('cms_privileges_roles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->boolean('is_visible')->nullable();
             $table->boolean('is_create')->nullable();
             $table->boolean('is_read')->nullable();
             $table->boolean('is_edit')->nullable();
             $table->boolean('is_delete')->nullable();
-            $table->integer('id_cms_privileges')->nullable();
-            $table->integer('id_cms_moduls')->nullable();
+            $table->bigInteger('id_cms_privileges')->nullable();
+            $table->bigInteger('id_cms_moduls')->nullable();
 
             $table->timestamps();
         });
